+++ vehicle/src/Vehicle/Backend/Loss.hs
@@ -3,7 +3,6 @@ module Vehicle.Backend.Loss
   )
 where
 
-import Control.Monad.Except (MonadError (..))
 import Data.Maybe (maybeToList)
 import Data.Proxy (Proxy (..))
 import Data.Set (Set)
@@ -58,6 +57,11 @@ convertDecls logicID logic requestedDecls = \case
     decls' <- addDeclEntryToContext normDecl $ convertDecls logicID logic requestedDecls decls
     return $ maybeToList maybeLossDecl ++ decls'
 
+shouldEmit :: Set Name -> GenericDecl expr -> Bool
+shouldEmit requestedDecls decl
+  | Set.null requestedDecls = isPropertyDecl decl
+  | otherwise = Set.member (nameOf (identifierOf decl)) requestedDecls
+
 convertDecl ::
   (MonadCompile m, MonadFreeContext Builtin m) =>
   DifferentiableLogicID ->
@@ -73,8 +77,7 @@ convertDecl logicID logic requestedDecls decl = do
           | isExternalResourceDecl decl -> Just <$> convertResourceDecl p ident sort typ
           | otherwise -> return Nothing
         DefFunction p ident ann typ expr
-          | isPropertyDecl decl || Set.member (nameOf ident) requestedDecls ->
-              Just <$> convertOutputDecl p ident ann typ expr
+          | shouldEmit requestedDecls decl -> Just <$> convertValueDecl p ident ann typ expr
           | otherwise -> return Nothing
         DefRecord {} -> return Nothing
 
@@ -91,7 +94,7 @@ convertResourceDecl p ident sort typ = do
   typ' <- convertDeclType typ
   return $ DefAbstract p ident sort typ'
 
-convertOutputDecl ::
+convertValueDecl ::
   (MonadLogic m) =>
   Provenance ->
   Identifier ->
@@ -99,77 +102,92 @@ convertOutputDecl ::
   VType Builtin ->
   Value Builtin ->
   m (Decl LossBuiltin)
-convertOutputDecl p ident ann typ value = do
-  leafType <- stripPiTypes typ
-  case toTypeValue leafType of
-    VBoolTensorType {} -> emit leafType
-    VRatTensorType {} -> emit leafType
-    VVectorType {} -> emit leafType
-    VBoolType -> emit leafType
-    VRatType -> emit leafType
-    _ ->
-      throwError $
-        UnimplementedFeature p $
-          "compiling declaration"
-            <+> quotePretty (nameOf ident)
-            <+> "as a loss output (supported leaf types: `Bool`, `Real`, `Tensor Bool _`, `Tensor Real _`, `Vector _ _`)"
-  where
-    emit leafType = do
-      lossType <- convertDeclType typ
-      lossValue <- convertFunction (convertMultiOutput leafType) value
-      let lossExpr = unnormalise 0 lossValue
-      return $ DefFunction p ident ann lossType lossExpr
-
-stripPiTypes :: (MonadLogic m) => VType Builtin -> m (VType Builtin)
-stripPiTypes typ = case toTypeValue typ of
-  VPiType binder closure -> do
-    body <- normaliseClosure binder closure
-    addNonTensorBinderToContext binder $ stripPiTypes body
-  _ -> return typ
+convertValueDecl p ident ann typ value = do
+  lossValue <- convertTypedValue p ident typ value
+  lossType <- convertDeclType typ
+  return $ DefFunction p ident ann lossType (unnormalise 0 lossValue)
 
 convertDeclType :: (MonadLogic m) => VType Builtin -> m (Type LossBuiltin)
 convertDeclType typ = unnormalise 0 <$> convertType typ
 
-convertMultiOutput :: (MonadLogic m) => VType Builtin -> Value Builtin -> m (Value LossBuiltin)
-convertMultiOutput typ = case toTypeValue typ of
-  VBoolTensorType _ds -> convertBoolTensorOutput
-  VRatTensorType _ds -> convertRatTensor
-  VVectorType tElem _d -> convertVectorOutput tElem
-  VBoolType -> convertBoolTensorOutput
-  VRatType -> convertRatTensor
-  _ -> unexpectedExprError currentPass "unsupported leaf type"
+convertTypedValue ::
+  (MonadLogic m) =>
+  Provenance ->
+  Identifier ->
+  VType Builtin ->
+  Value Builtin ->
+  m (Value LossBuiltin)
+convertTypedValue p ident typ value = case toTypeValue typ of
+  VPiType binder closure -> case value of
+    VLam vBinder vClosure -> do
+      vBinder' <- traverse convertType vBinder
+      let convertBody v = do
+            body <- normaliseClosure binder closure
+            addNonTensorBinderToContext binder $ convertTypedValue p ident body v
+      vClosure' <- convertClosure convertBody vBinder vClosure
+      return $ VLam vBinder' vClosure'
+    -- Non-Lam at a Pi position: walk past the binder and dispatch at the leaf.
+    _ -> do
+      body <- normaliseClosure binder closure
+      addNonTensorBinderToContext binder $ convertTypedValue p ident body value
+  VBoolType -> convertBoolTensorValue value
+  VRatType -> convertRatTensor value
+  VNatType -> convertNatValue value
+  VIndexType _ -> convertIndexValue value
+  VBoolTensorType _ -> convertBoolTensorValue value
+  VRatTensorType _ -> convertRatTensor value
+  VNatTensorType _ -> convertNatTensor value
+  VIndexTensorType _ _ -> convertIndexTensor value
+  VVectorType tElem _d -> convertVectorValue p ident tElem value
+  VListType tElem -> convertListValue (convertTypedValue p ident tElem) value
+  VUnitType {} -> unsupportedOperation "Unit-typed declarations"
+  VFreeTypeVar {} -> unexpectedExprError currentPass "free type variable in decl type"
+  VBoundTypeVar {} -> unexpectedExprError currentPass "bound type variable in decl type"
 
-convertVectorOutput :: (MonadLogic m) => VType Builtin -> Value Builtin -> m (Value LossBuiltin)
-convertVectorOutput typ value = do
-  let dims = getVectorDims typ
+convertVectorValue ::
+  (MonadLogic m) =>
+  Provenance ->
+  Identifier ->
+  VType Builtin ->
+  Value Builtin ->
+  m (Value LossBuiltin)
+convertVectorValue p ident elemType value = do
+  let dims = getVectorDims elemType
+  let convertElem = convertTypedValue p ident elemType
   case toVectorValue value of
     VVectorBoundVar lv spine -> convertBoundVar lv spine
-    VVectorDataset ident -> return $ VFreeVar ident []
-    VVectorLiteral args -> convertVecLiteralArgs (convertMultiOutput typ) (IBoolType, dims) args
+    VVectorDataset name -> return $ VFreeVar name []
+    VVectorLiteral args -> convertVecLiteralArgs convertElem (IBoolType, dims) args
     VVectorIf args -> convertIf args
-    VVectorForeach args -> convertVecForeachArgs (convertMultiOutput typ) (IBoolType, dims) args
+    VVectorForeach args -> convertVecForeachArgs convertElem (IBoolType, dims) args
 
-convertBoolTensorOutput :: (MonadLogic m) => Value Builtin -> m (Value LossBuiltin)
-convertBoolTensorOutput value = case toBoolTensorValue value of
+getVectorDims :: VType Builtin -> VDims Builtin
+getVectorDims typ = case toTypeValue typ of
+  VBoolTensorType ds -> ds
+  VRatTensorType ds -> ds
+  VNatTensorType ds -> ds
+  VIndexTensorType _ ds -> ds
+  VVectorType t d -> IDimCons d (getVectorDims t)
+  _ -> developerError "non-tensor element type in vector unwrap"
+
+-- `LossCompilation.convertBoolTensor` throws on `VBoolTensorQuantifyRat`
+-- because its callers nest inside contexts where quantifiers can't occur.
+-- A top-level decl body can have them, so handle that one arm here.
+convertBoolTensorValue :: (MonadLogic m) => Value Builtin -> m (Value LossBuiltin)
+convertBoolTensorValue value = case toBoolTensorValue value of
   VBoolTensorLiteral bs -> convertBoolTensorLiteral bs
-  VBoolConstTensor args -> convertConstTensor convertBoolTensorOutput args
-  VBoolStackTensor args -> convertStackTensor convertBoolTensorOutput args
-  VBoolTensorAnd args -> convertAnd =<< convertTensorOp2 convertBoolTensorOutput args
-  VBoolTensorOr args -> convertOr =<< convertTensorOp2 convertBoolTensorOutput args
-  VBoolTensorNot args -> convertNot =<< convertTensorOp1 convertBoolTensorOutput args
+  VBoolConstTensor args -> convertConstTensor convertBoolTensorValue args
+  VBoolStackTensor args -> convertStackTensor convertBoolTensorValue args
+  VBoolTensorAnd args -> convertAnd =<< convertTensorOp2 convertBoolTensorValue args
+  VBoolTensorOr args -> convertOr =<< convertTensorOp2 convertBoolTensorValue args
+  VBoolTensorNot args -> convertNot =<< convertTensorOp1 convertBoolTensorValue args
   VBoolTensorCompareNat args -> convertNatComparison args
   VBoolTensorCompareIndex args -> convertIndexComparison args
   VBoolTensorCompareRatPointwise args -> convertRatTensorPointwiseComparison args
   VBoolTensorCompareRatReduced args -> convertRatTensorReducedComparison args
   VBoolTensorQuantifyRat args -> compileQuantifier args
-  VBoolTensorReduceAnd args -> convertReduceAnd =<< convertTensorReduction convertBoolTensorOutput args
-  VBoolTensorReduceOr args -> convertReduceOr =<< convertTensorReduction convertBoolTensorOutput args
+  VBoolTensorReduceAnd args -> convertReduceAnd =<< convertTensorReduction convertBoolTensorValue args
+  VBoolTensorReduceOr args -> convertReduceOr =<< convertTensorReduction convertBoolTensorValue args
   VBoolTensorBoolIf args -> convertIf args
-  VBoolTensorAt args -> convertAtTensor convertBoolTensorOutput args
-  VBoolTensorForeach args -> convertForeachTensor convertBoolTensorOutput args
-
-getVectorDims :: VType Builtin -> VDims Builtin
-getVectorDims typ = case toTypeValue typ of
-  VBoolTensorType ds -> ds
-  VVectorType t d -> IDimCons d (getVectorDims t)
-  _ -> developerError "Impossible property type"
+  VBoolTensorAt args -> convertAtTensor convertBoolTensorValue args
+  VBoolTensorForeach args -> convertForeachTensor convertBoolTensorValue args